Система начисления рейтинга (+1 +2 +3 за правильный ответ на вопрос, в зависимости от его сложности; -1 за неправильный ответ; рейтинг не может быть < 0) Небольшие фиксы (unused imports, long -> int у рейтинга у юзера и прочее...)

+++ src/main/java/itmo/help_interview/bot/entity/User.java
@@ -30,7 +30,7 @@ public class User {
 	private boolean isBanned;
 
 	@Column(name = "rating", nullable = false)
-	private long rating;
+	private int rating;
 
 	@ManyToMany(fetch = FetchType.EAGER)
 	@JoinTable(
